adapt to new backend API
