+++ src/app/organizations/organization-details/manage-projects/manage-projects.component.ts
@@ -2,7 +2,7 @@ import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
 import { OrganizationService } from '../../../shared/services/project/organization.service';
 import { displayBackendErrorRx } from '../../../utilities/error-handler';
-import { Project, ProjectService } from '../../../shared/services/project/project.service';
+import { Project, ProjectService, ProjectWallet } from '../../../shared/services/project/project.service';
 import { BehaviorSubject, Observable } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
 import { MiddlewareService } from '../../../shared/services/middleware/middleware.service';
@@ -16,22 +16,24 @@ export class ManageProjectsComponent {
     @Input() groupID: string;
 
     refreshProjectsSubject = new BehaviorSubject<void>(null);
-    projects$: Observable<Project[]>;
+    projectsWallets$: Observable<ProjectWallet[]>;
 
     constructor(private router: Router,
                 private orgService: OrganizationService,
                 private middlewareService: MiddlewareService,
                 private projectService: ProjectService) {
-        this.projects$ = this.refreshProjectsSubject.pipe(
+        this.projectsWallets$ = this.refreshProjectsSubject.pipe(
             switchMap(() => this.orgService.getAllProjectsForOrganization(this.groupID)
                 .pipe(displayBackendErrorRx())),
-            map(res => res.projects)
+            map(res => res.projects_wallets)
         );
     }
 
-    toggleProject(uuid: string, projects: Project[]) {
+    toggleProject(uuid: string, projectsWallets: ProjectWallet[]) {
         return () => {
-            const project = projects.find(proj => proj.uuid === uuid);
+            const project = projectsWallets
+                .map(pw => pw.project)
+                .find(proj => proj.uuid === uuid);
             return this.projectService.updateProject(project.uuid, {
                 active: !project.active
             }).pipe(
@@ -41,7 +43,7 @@ export class ManageProjectsComponent {
         };
     }
 
-    projectWallet(hash: string) {
+    getWalletInfo(hash: string) {
         return this.middlewareService.getProjectWalletInfoCached(hash);
     }
 }
