+++ src/app/organizations/organization-details/organization-details.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { SpinnerUtil } from 'src/app/utilities/spinner-utilities';
 import { displayBackendErrorRx } from 'src/app/utilities/error-handler';
-import { WalletService } from '../../shared/services/wallet/wallet.service';
-import { WalletDetails } from '../../shared/services/wallet/wallet-cooperative/wallet-cooperative-wallet.service';
+import { Wallet, WalletService } from '../../shared/services/wallet/wallet.service';
 import { Organization, OrganizationMember, OrganizationService } from '../../shared/services/project/organization.service';
 import { BehaviorSubject, EMPTY, Observable, of } from 'rxjs';
 import { catchError, finalize, map, switchMap, tap } from 'rxjs/operators';
@@ -22,7 +21,7 @@ export class OrganizationDetailsComponent implements OnInit {
     refreshOrgMembersSubject = new BehaviorSubject<void>(null);
 
     organization$: Observable<Organization>;
-    orgWallet$: Observable<WalletDetails>;
+    orgWallet$: Observable<Wallet>;
     orgMembers$: Observable<OrganizationMember[]>;
 
     inviteForm: FormGroup;
@@ -144,7 +143,7 @@ export class OrganizationDetailsComponent implements OnInit {
         ).subscribe();
     }
 
-    isWalletVerified(orgWallet: WalletDetails) {
+    isWalletVerified(orgWallet: Wallet) {
         return !!orgWallet && !!orgWallet?.hash;
     }
