+++ src/app/shared/services/wallet/portfolio.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BackendHttpClient } from '../backend-http-client.service';
 import { Project } from '../project/project.service';
-import { TransactionInfo } from './wallet-cooperative/wallet-cooperative-wallet.service';
+import { TransactionInfo } from './wallet.service';
 
 @Injectable({
     providedIn: 'root'
