+++ src/app/shared/services/wallet/revenue-share.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { BackendHttpClient } from '../backend-http-client.service';
-import { TransactionInfo } from './wallet-cooperative/wallet-cooperative-wallet.service';
+import { TransactionInfo } from './wallet.service';
 
 @Injectable({
     providedIn: 'root'
